Some little improvements on layer3
Improve a little code:
- timeout is casted to (uint16_t), coz its bigger than 255
- within the receiveData USART::transmit is done after LAYER2::transmit,
  due to the fact that usart will block and this will delay the layer2 transmit
- within the receiveData its checked if the transmitBuffer is empty,
  otherwise this could cause the assert(!isEmpty()) to kill the running program

+++ src/radio/layer3/layer3.h
@@ -11,7 +11,7 @@
 #define PACKET_CODE_DATA 0
 #define PACKET_CODE_ACK 1
 
-#define TIMEOUT 400
+#define TIMEOUT (uint16_t) 400
 #define TIMEOUT_MAX_RESEND 3
 
 #include <stdint.h>
